microshif: reorganize config-svc in modules and templates
