+++ controller/deploy/microshift-bootc/config-svc/__init__.py
@@ -1,0 +1,4 @@
+"""Jumpstarter Configuration Web UI package."""
+
+__version__ = "1.0.0"
+
