+++ controller/deploy/microshift-bootc/config-svc/config-svc.service
@@ -7,7 +7,7 @@ Before=getty@.service systemd-user-sessions.service
 
 [Service]
 Type=simple
-ExecStart=/usr/bin/python3 /usr/local/bin/config-svc
+ExecStart=/usr/local/bin/config-svc
 Restart=on-failure
 RestartSec=5
 Environment="PORT=8880"
