+++ controller/deploy/microshift-bootc/config-svc/templates/password_required.html
@@ -1,0 +1,134 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Password Change Required - Jumpstarter</title>
+    <link href="https://jumpstarter.dev/_static/favicon.png" rel="shortcut icon">
+    <link rel="stylesheet" href="/static/styles.css">
+</head>
+<body>
+    <div class="container">
+        <div class="banner">
+            <div class="banner-text">Jumpstarter Controller Community Edition • Powered by MicroShift</div>
+            <div class="logos">
+                <a href="https://jumpstarter.dev/" target="_blank" rel="noopener noreferrer" class="logo-link" title="Jumpstarter - Remote device management and testing">
+                    <img src="https://jumpstarter.dev/main/_static/logo-light-theme.svg" alt="Jumpstarter" class="jumpstarter-logo" />
+                </a>
+                <a href="https://microshift.io/" target="_blank" rel="noopener noreferrer" class="logo-link" title="MicroShift - Optimized OpenShift for the device edge">
+                    <img src="https://microshift.io/images/microshift_logo_white.png" alt="MicroShift" class="microshift-logo" />
+                </a>
+            </div>
+        </div>
+        
+        <div class="content-area">
+            <h1>Security Setup Required</h1>
+            
+            {% for msg in messages %}
+            <div class="message {{ msg.type }}">{{ msg.text }}</div>
+            {% endfor %}
+            
+            <div class="warning-box">
+                <h2>⚠️ Default Password Detected</h2>
+                <p>You are using the default password. For security reasons, you must change the root password before accessing the configuration interface.</p>
+            </div>
+            
+            <form id="password-change-form">
+                <div id="password-messages-container"></div>
+                <div class="form-group">
+                    <label for="newPassword">New Root Password *</label>
+                    <input type="password" id="newPassword" name="newPassword" minlength="8" autofocus>
+                    <div class="hint">Minimum 8 characters (required to change from default password)</div>
+                </div>
+                <div class="form-group">
+                    <label for="confirmPassword">Confirm Password *</label>
+                    <input type="password" id="confirmPassword" name="confirmPassword" minlength="8">
+                    <div class="hint">Re-enter your new password</div>
+                </div>
+                <div class="form-group">
+                    <label for="sshKeys">SSH Authorized Keys (Optional)</label>
+                    <textarea id="sshKeys" name="sshKeys" rows="6">{{ ssh_keys }}</textarea>
+                    <div class="hint">One SSH public key per line. Leave empty to clear existing keys.</div>
+                </div>
+                <button type="submit" id="password-submit-btn">Change Password & Continue</button>
+            </form>
+            <script>
+                (function() {
+                    const form = document.getElementById('password-change-form');
+                    const messagesContainer = document.getElementById('password-messages-container');
+                    const submitBtn = document.getElementById('password-submit-btn');
+                    
+                    form.addEventListener('submit', function(e) {
+                        e.preventDefault();
+                        
+                        const data = {
+                            newPassword: document.getElementById('newPassword').value,
+                            confirmPassword: document.getElementById('confirmPassword').value,
+                            sshKeys: document.getElementById('sshKeys').value
+                        };
+                        
+                        const originalText = submitBtn.textContent;
+                        submitBtn.disabled = true;
+                        submitBtn.textContent = 'Processing...';
+                        
+                        // Clear previous messages
+                        messagesContainer.innerHTML = '';
+                        
+                        fetch('/api/change-password', {
+                            method: 'POST',
+                            headers: {
+                                'Content-Type': 'application/json',
+                            },
+                            credentials: 'same-origin',
+                            body: JSON.stringify(data)
+                        })
+                        .then(response => response.json())
+                        .then(result => {
+                            // Display messages
+                            result.messages.forEach(msg => {
+                                const messageDiv = document.createElement('div');
+                                messageDiv.className = `message ${msg.type}`;
+                                messageDiv.textContent = msg.text;
+                                messagesContainer.appendChild(messageDiv);
+                            });
+                            
+                            // Update SSH keys textarea if they were updated
+                            if (result.ssh_updated && result.ssh_keys !== undefined) {
+                                document.getElementById('sshKeys').value = result.ssh_keys;
+                            }
+                            
+                            // If password was changed from default, trigger redirect
+                            if (result.requires_redirect) {
+                                // Hide form and show redirect message
+                                form.style.display = 'none';
+                                messagesContainer.innerHTML = '<div class="message success">Password changed successfully! Redirecting to login with your new password...</div>';
+                                
+                                setTimeout(function() {
+                                    fetch('/logout', {
+                                        method: 'GET',
+                                        headers: {
+                                            'Authorization': 'Basic ' + btoa('logout:logout')
+                                        }
+                                    }).finally(function() {
+                                        window.location.href = '/';
+                                    });
+                                }, 2000);
+                            } else if (result.success) {
+                                // Scroll to messages
+                                messagesContainer.scrollIntoView({ behavior: 'smooth', block: 'nearest' });
+                            }
+                        })
+                        .catch(error => {
+                            messagesContainer.innerHTML = '<div class="message error">Failed to update: ' + error.message + '</div>';
+                        })
+                        .finally(() => {
+                            submitBtn.disabled = false;
+                            submitBtn.textContent = originalText;
+                        });
+                    });
+                })();
+            </script>
+        </div>
+    </div>
+</body>
+</html>
