+++ controller/deploy/microshift-bootc/config-svc/update-banner.sh
@@ -3,34 +3,30 @@
 
 import sys
 import os
-import importlib.util
-import importlib.machinery
 
-def update_login_banner():
-config_svc_path = '/usr/local/bin/config-svc'
-if not os.path.exists(config_svc_path):
-    print(f"Error: {config_svc_path} does not exist", file=sys.stderr)
-    sys.exit(1)
+# Add config-svc module directory to Python path
+sys.path.insert(0, '/usr/local/lib/config-svc')
 
-# Try to create spec with explicit loader for files without .py extension
 try:
-    # Use SourceFileLoader explicitly for files without .py extension
-    loader = importlib.machinery.SourceFileLoader('config_svc', config_svc_path)
-    spec = importlib.util.spec_from_loader('config_svc', loader)
+    # Import the system module which contains update_login_banner()
+    from system import update_login_banner
     
-    if spec is None:
-        print(f"Error: Failed to create spec for {config_svc_path}", file=sys.stderr)
+    # Call the update function
+    success, message = update_login_banner()
+    
+    if not success:
+        print(f"Warning: Failed to update login banner: {message}", file=sys.stderr)
         sys.exit(1)
     
-    if spec.loader is None:
-        print(f"Error: Failed to get loader for {config_svc_path}", file=sys.stderr)
-        sys.exit(1)
+    print("Login banner updated successfully", file=sys.stdout)
+    sys.exit(0)
     
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    module.update_login_banner()
+except ImportError as e:
+    print(f"Error: Failed to import system module: {e}", file=sys.stderr)
+    print("Make sure config-svc modules are installed at /usr/local/lib/config-svc/", file=sys.stderr)
+    sys.exit(1)
 except Exception as e:
-    print(f"Error loading or executing {config_svc_path}: {e}", file=sys.stderr)
+    print(f"Error updating login banner: {e}", file=sys.stderr)
     sys.exit(1)
